Fix the portalSpectatorDelay module's descriptions. Also move validators to their own package.
