+++ src/main/java/carpetaddonsnotfound/CarpetAddonsNotFoundSettings.java
@@ -1,14 +1,13 @@
 package carpetaddonsnotfound;
 
-import carpet.api.settings.CarpetRule;
 import carpet.api.settings.Rule;
-import carpet.api.settings.Validator;
-import net.minecraft.server.command.ServerCommandSource;
+import carpetaddonsnotfound.validators.OneHourMaxDelayLimit;
 
 import static carpet.api.settings.RuleCategory.*;
 
 public class CarpetAddonsNotFoundSettings {
   private static final String CARPET_ADDONS_NOT_FOUND = "carpet-addons-not-found";
+  private static final String SPECTATOR = "spectator";
 
   @Rule(categories = { FEATURE, CREATIVE, CARPET_ADDONS_NOT_FOUND })
   public static boolean creativePlayerOneHitKill = false;
@@ -66,7 +65,7 @@ public class CarpetAddonsNotFoundSettings {
 
   @Rule(
           options = { "1", "40", "80", "72000" },
-          categories = { CREATIVE, CARPET_ADDONS_NOT_FOUND },
+          categories = { SPECTATOR, CARPET_ADDONS_NOT_FOUND },
           strict = false,
           validators = OneHourMaxDelayLimit.class
   )
@@ -75,7 +74,7 @@ public class CarpetAddonsNotFoundSettings {
   @Rule(categories = { FEATURE, SURVIVAL, CARPET_ADDONS_NOT_FOUND })
   public static ReplaceableFlowersOptions replaceableFlowers = ReplaceableFlowersOptions.FALSE;
 
-  @Rule(categories = { FEATURE, CREATIVE, CARPET_ADDONS_NOT_FOUND })
+  @Rule(categories = { SPECTATOR, CARPET_ADDONS_NOT_FOUND })
   public static boolean spawnEggsSpawnMobsWithNoAI = false;
 
   @Rule(categories = { FEATURE, EXPERIMENTAL, CARPET_ADDONS_NOT_FOUND })
@@ -94,19 +93,5 @@ public class CarpetAddonsNotFoundSettings {
     FALSE,
     ONE_TALL_FLOWERS(),
     ALL_FLOWERS()
-  }
-
-
-  private static class OneHourMaxDelayLimit extends Validator<Integer> {
-    @Override
-    public Integer validate(ServerCommandSource source, CarpetRule<Integer> currentRule, Integer newValue,
-                            String string) {
-      return (newValue > 0 && newValue <= 72000) ? newValue : null;
-    }
-
-    @Override
-    public String description() {
-      return "You must choose a value from 1 to 72000";
-    }
   }
 }
